Rearrange grid, card and index page components
It didn't make sense to call the restaurants grid component "desktop"
because it's the same component for mobile.

The "RestaurantIndex" component name wasn't clear enough,
so I appended "Page" to it: "RestaurantIndexPage".

I moved the RestaurantGridPage to its own file for organization.
