+++ src/components/RestaurantGrid.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import { RestaurantGridDesktop } from './RestaurantGridDesktop';
+import { RestaurantGrid } from './RestaurantGrid';
 
 import {
   LongName as LongNameCard,
@@ -9,11 +9,11 @@ import {
 } from './RestaurantCardDesktop.stories'
 
 export default {
-  title: 'Components/RestaurantGridDesktop',
-  component: RestaurantGridDesktop,
-} as ComponentMeta<typeof RestaurantGridDesktop>;
+  title: 'Components/RestaurantGrid',
+  component: RestaurantGrid,
+} as ComponentMeta<typeof RestaurantGrid>;
 
-const Template: ComponentStory<typeof RestaurantGridDesktop> = (args) => <RestaurantGridDesktop {...args} />;
+const Template: ComponentStory<typeof RestaurantGrid> = (args) => <RestaurantGrid {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {
