+++ src/pages/RestaurantIndexPage.tsx
@@ -4,11 +4,10 @@ import { Divider } from "../components/Divider";
 import { Heading1 } from "../components/Heading1";
 import { Heading2 } from "../components/Heading2";
 import { Layout } from "../components/Layout";
-import { RestaurantGridDesktopContainer } from "../components/RestaurantCardsContainer";
-import { RestaurantGridDesktop } from "../components/RestaurantGridDesktop";
+import { RestaurantGridContainer } from "../components/RestaurantGridContainer";
 import { Subheading } from "../components/Subheading";
 
-export const RestaurantIndex: FC = () => {
+export const RestaurantIndexPage: FC = () => {
   return (
     <Layout>
       <HeroMini>
@@ -17,7 +16,7 @@ export const RestaurantIndex: FC = () => {
       </HeroMini>
       <Divider />
       <StyledHeading2>All Restaurants</StyledHeading2>
-      <RestaurantGridDesktopContainer />
+      <RestaurantGridContainer />
     </Layout>
   )
 }
